+++ index.js
@@ -328,6 +328,11 @@ class WebdriverAjax {
           body: parseBody(req.body),
           statusCode: req.statusCode,
         };
+        if (!req.headers) {
+          console.warn(
+            `${transformed.method} request to ${req.url} (HTTP ${req.statusCode}) had no response headers!`
+          );
+        }
       }
       return transformed;
     }
@@ -344,6 +349,9 @@ class WebdriverAjax {
     // (best effort compliance with RFC)
     function parseResponseHeaders(rawHeader) {
       const headers = {};
+      if (!rawHeader) {
+        return headers;
+      }
       const lines = rawHeader.trim().split(/(?:\r?\n)+/);
       for (const line of lines) {
         const parts = line.split(/(?<=^[^:]*):/);